realtime_ingestion/poller/requirements.txt
# Poller dependencies - fetch, parse protobuf extensions, publish
google-cloud-pubsub>=2.18.0
requests>=2.31.0
protobuf>=6.33.0
